+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Features:
  * Igneous Miner now has a small mining cooldown
+ * Igneous Crafter now has a crafting preview
 
 Fixes:
  * Igneous Crafter now supports recipes with multiple variants (the one in the template is used)